requirements.txt
keyboard>=0.13.5
pyperclip>=1.8.2
Pillow>=9.0.0
pystray>=0.19.0
